code review without those commented by me
